refactor: keep readMdMdxFiles in server.ts
